implement GameDAO, and tests

+++ server/src/main/java/dataaccess/SQLAuthDAO.java
@@ -31,6 +31,9 @@ public class SQLAuthDAO implements AuthDAO {
 
     @Override
     public void deleteAuth(String authToken) throws DataAccessException {
+        if(getAuth(authToken) == null){
+            throw new DataAccessException("Can't Delete what doesn't exist");
+        }
         var statement = "DELETE FROM auth WHERE  authToken=?";
         executeUpdate(statement, authToken);
     }
